MAINT: Move HappiChannel
